\name{base64-package}
\alias{base64-package}
\alias{base64}
\docType{package}
\title{
Base64 encoder/decoder
}
\description{
Base64 encoder/decoder
}
\details{
\tabular{ll}{
Package: \tab base64\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-04-03\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Romain Francois, based on http://base64.sourceforge.net/

Maintainer: Romain Francois <romain@r-enthusiasts.com>
}
\keyword{ package }

